I add end of trip greeting and fix showDialog parking

+++ lib/roles/driver/infraestructure/entities/driver_trip_state.dart
@@ -112,7 +112,7 @@ class Observations {
 
 class Parking {
   int id;
-  int amount;
+  double amount;
   String name;
   int tripId;
